Bot ultime/requirements.txt
# Dependencies for Lighter DEX (Python 3.9+)
lighter-sdk
aiohttp

# Dependencies for Paradex DEX (Python 3.11)
paradex-py
starknet-py

# Dependencies communes
supabase
psycopg2-binary
python-dotenv
numpy

# Web server
# (http.server is built-in)
